feat: 知识库语义化检索 + 全面代码质量优化
## P0-3 知识库语义化（S1-S4）
- EmbeddingProvider/SemanticIndex 协议 + IndexHit/IndexFilter 类型
- KeywordEmbedder（降级 mock）+ InMemoryVectorIndex（vault 扫描）
- MLXEmbeddingProvider（bge-m3 原生 mlx-swift-lm + HF 镜像下载）
- SQLiteVectorIndex（对接 XiaoNuo schema，零拷贝 vDSP + 两阶段检索）
- RuleEngine 六步管道重构（概念提取→索引反查→语义兜底→全文→跨源→组装）

## 代码质量审查优化（Phase 0-3，19项全完成）
- Phase 0: SQLite 零拷贝 + 测试修复(142→0) + weak self + static var注入
- Phase 1: CompactionWatermark增量token + AgentMetrics单锁 + SwiftLint 73→0
- Phase 2: waitUntilReady + 一键入口 + 配置统一 + 工具stub→注入式实现
- Phase 3: actor迁移(3文件) + ToolDispatch拆分(859→4文件) + try?审计
  + JSONValue类型安全([String:Any]→[String:JSONValue]) + API文档80%+

+++ Packages/YunPatCore/Tests/YunPatCoreTests/CostTrackerTests.swift
@@ -21,7 +21,7 @@ final class CostTrackerTests: XCTestCase {
 
     func test_nilUsage_ignored_returnsZero() async {
         let tracker: CostTracker = CostTracker(maxBudgetTokens: 10_000)
-        let delta: Int = await tracker.record(usage: nil, model: "test")
+        let delta: Double = await tracker.record(usage: nil, model: "test")
         XCTAssertEqual(delta, 0)
         let snap: CostSnapshot = await tracker.snapshot
         XCTAssertTrue(snap.isZero)
